Fix parent_id column type to match servers.id int unsigned

+++ database/migrations/2026_07_23_000000_add_server_splitting_to_servers_table.php
@@ -9,7 +9,7 @@ return new class extends Migration
     public function up(): void
     {
         Schema::table('servers', function (Blueprint $table) {
-            $table->unsignedBigInteger('parent_id')->nullable()->index()->after('owner_id');
+            $table->unsignedInteger('parent_id')->nullable()->index()->after('owner_id');
             $table->unsignedInteger('split_limit')->default(0)->after('backup_limit');
 
             $table->foreign('parent_id')->references('id')->on('servers')->nullOnDelete();
